feat: gpio_monitor external sensor componenet
gpio_monitor sensor detects a weird issue with an ESP8266 with
WiFi and ADC running perfectly but GPIO subsystem unusuable.

Every minute, it configures an unused GPIO passed as parameter
as an output, writes 1, redefines it as input and reads it. Then,
it repeats with with a 0.

If we do not read the value written to the pin it means that pin
is connected to something or GPIO subsystem is hang up.