feat: delta_sensor component to replace custom components
delta_sensor allows to calculate the difference between successive
states of a text sensor divided by the time in seconds between
the state updates.

This allows to convert energy sensors to power sensors with larger
precision than regular template sensors that use 32-bit floating
point numbers, but can be used for any other purpose.

It replaces custom components in the my electric counter, which
are deprecated and will be removed soon.